functions/recognizer/recognizer__violence/requirements.txt
fastapi
uvicorn
Pillow
numpy
redis
tensorflow
boto3